Combine native and native-view into native-common
