+++ packages/create-react-native-library/templates/native-common/android/src/main/AndroidManifest.xml
