+++ packages/create-react-native-library/templates/native-common/ios/{%- project.name %}.xcodeproj/project.xcworkspace/contents.xcworkspacedata
