+++ packages/create-react-native-library/templates/native-common/{%- project.podspec %}.podspec
