--- packages/create-react-native-library/templates/native-library/src/index.tsx
@@ -1,9 +1,0 @@
-import { NativeModules } from 'react-native';
-
-type <%- project.name %>Type = {
-  multiply(a: number, b: number): Promise<number>;
-};
-
-const { <%- project.name %> } = NativeModules;
-
-export default <%- project.name %> as <%- project.name %>Type;
