--- packages/create-react-native-library/templates/native-view-library/android/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.<%- project.package %>">
-
-</manifest>
