--- packages/create-react-native-library/templates/native-view-library/ios/{%- project.name %}.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-</Workspace>
