--- packages/create-react-native-library/templates/native-view-library/{%- project.podspec %}.podspec
@@ -1,25 +1,0 @@
-require "json"
-
-package = JSON.parse(File.read(File.join(__dir__, "package.json")))
-
-Pod::Spec.new do |s|
-  s.name         = "<%- project.podspec %>"
-  s.version      = package["version"]
-  s.summary      = package["description"]
-  s.homepage     = package["homepage"]
-  s.license      = package["license"]
-  s.authors      = package["author"]
-
-  s.platforms    = { :ios => "10.0" }
-  s.source       = { :git => "<%- repo %>.git", :tag => "#{s.version}" }
-
-<% if (project.cpp) { -%>
-  s.source_files = "ios/**/*.{h,m,mm}", "cpp/**/*.{h,cpp}"
-<% } else if (project.swift) { -%>
-  s.source_files = "ios/**/*.{h,m,mm,swift}"
-<% } else { -%>
-  s.source_files = "ios/**/*.{h,m,mm}"
-<% } -%>
-
-  s.dependency "React-Core"
-end
